+++ R/release.R
@@ -1,0 +1,5 @@
+release_questions <- function() {
+  c(
+    "Did you re-build the vignettes using `vignettes/precompile.R`?"
+  )
+}