+++ NEWS.md
@@ -1,5 +1,7 @@
 # rphylopic (development version)
 
+* vignettes are now precompiled
+
 # rphylopic 1.2.1
 
 * updated rphylopic to work with grImport2 version 0.3.0 and rsvg version 2.6.0